+++ src/main/java/io/dataspaceconnector/model/config/DatabaseConstants.java
@@ -26,6 +26,11 @@ public final class DatabaseConstants {
     public static final int URI_COLUMN_LENGTH = 2048;
 
     /**
+     * The maximum length of database columns containing descriptions.
+     */
+    public static final int DESCRIPTION_COLUMN_LENGTH = 4096;
+
+    /**
      * Private constructor.
      */
     private DatabaseConstants() { }